requirements.txt
streamlit
streamlit_webrtc
numpy
torch
pandas
torchvision
opencv-python
matplotlib
ipython
psutil
pyyaml
tqdm
seaborn
mysql-connector-python==8.0.32
plotly==5.14.1